change reminders to new layout
