+++ app/views/reminders/_header.html.haml
@@ -1,0 +1,1 @@
+%h1.ui.header= title
