+++ app/views/reminders/_menu.html.haml
@@ -1,0 +1,1 @@
+= active_link_to t('reminders.experts.menu.experts'), reminders_experts_path, { class: 'item' }
