+++ app/views/reminders/experts/_menu.html.haml
@@ -1,0 +1,4 @@
+= active_link_to(t('.summary'), reminders_expert_path(expert), active: ['reminders/experts': :show], class: 'item')
+= active_link_to(t('needs.index.needs_quo'), needs_reminders_expert_path(expert), class: 'item')
+= active_link_to(t('needs.index.needs_taking_care'), needs_taking_care_reminders_expert_path(expert), class: 'item')
+= active_link_to(t('needs.index.needs_others_taking_care'), needs_taking_care_by_others_reminders_expert_path(expert), class: 'item')
