+++ app/views/reminders/experts/index.haml
@@ -1,9 +1,9 @@
 - meta title: t('.title')
 
-%h1.ui.header
-  = t('.title')
+- content_for :header, render('reminders/header', title: t('.title'))
+- content_for :menu, render('reminders/menu')
 
-= form_with url: reminders_path, method: :get, skip_enforcing_utf8: true, class: 'ui large form' do
+= form_with url: reminders_experts_path, method: :get, skip_enforcing_utf8: true, class: 'ui large form' do
   - options = { include_blank: t('all_choices'), class: 'ui search selection dropdown', onchange: 'this.form.submit();' }
   .inline.field
     = label_tag(t('activerecord.models.territory.one'))
