--- app/views/reminders/show.haml
@@ -1,17 +1,0 @@
-- title = t('.title', full_name: @expert.full_name)
-
-- meta title: title
-
-%h1.ui.header
-  = title
-.ui.segments.shadow-less
-  .ui.segment
-    = render 'expert_contents', expert: @expert
-    %div{ id: "expert_reminders_notes-#{@expert.id}" }
-      = render 'expert_reminders_notes', expert: @expert
-
-= render 'reminders_tabs', expert: @expert, cached: true
-
-%p= page_entries_info @needs
-
-= render 'needs/received_needs', received_needs: @needs, highlighted_expert: @expert, cached: true
