+++ config/routes.rb
@@ -163,11 +163,14 @@ Rails.application.routes.draw do
   resources :matches, only: %i[update]
   resources :feedbacks, only: %i[create destroy]
 
-  resources :reminders, only: %i[index show], path: 'relances' do
-    member do
-      post :reminders_notes
-      get :needs_taking_care
-      get :needs_taking_care_by_others
+  namespace :reminders, path: 'relances' do
+    resources :experts, only: %i[index show] do
+      member do
+        post :reminders_notes
+        get :needs
+        get :needs_taking_care
+        get :needs_taking_care_by_others
+      end
     end
   end
